Updated services and repository to utilise service bus queues for deleting packages and photos

+++ localsound.backend/AzureFunctions/localsound.CoreUpdates/Repository/BlobRepository.cs
@@ -30,6 +30,12 @@ namespace localsound.CoreUpdates.Repository
 
                 var blobClient = blobContainerClient.GetBlobClient(blobPath);
 
+                // If it doesnt exist then we dont need to delete it
+                if (!(await blobClient.ExistsAsync()))
+                {
+                    return true;
+                }
+
                 // Delete the file if it already exists
                 var result = await blobClient.DeleteIfExistsAsync();
